# Finalizing CRUD Website

+++ backend/controller/product.controller.js
@@ -46,11 +46,15 @@ export const updateProducts = async (req,res) =>{
 export const deleteProducts = async (req,res) => {
     const {id} = req.params;
     
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success: false, message: "Invalid Product Id"})
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true,  message: "Product Deleted"});
     } catch (error) {
         console.log("error in deleting products:", error.message);
-        return res.status(404).json({ success: false, message: "Product Not Found" });
+        return res.status(500).json({ success: false, message: "Server Error" });
     }
 }
